Add DecimalValue (neater rounding then Floats)
  - Update host page to add MetricPrefix and convert unit
    milli or micro units if appropriate.

+++ argux_server/monitors/ICMPMonitor.py
@@ -6,6 +6,8 @@ import platform
 import re
 import subprocess
 
+from decimal import Decimal
+
 from datetime import datetime
 
 from .AbstractMonitor import AbstractMonitor
@@ -24,7 +26,7 @@ def parse_freebsd(monitor, output):
     for row in output.split('\n'):
         m = re.search('^round-trip min/avg/max/stddev = (?P<min>[0-9]+(?:\.[0-9]+)?).*', row)
         if (m):
-            ret_val = str(float(m.group(1))/1000)
+            ret_val = str(Decimal(m.group(1))/1000)
     return ret_val
 
 def parse_linux(monitor, output):
@@ -41,7 +43,7 @@ def parse_linux(monitor, output):
     for row in output.split('\n'):
         m = re.search('^rtt min/avg/max/mdev = (?P<min>[0-9]+(?:\.[0-9]+)?).*', row)
         if (m):
-            ret_val = str(float(m.group(1))/1000)
+            ret_val = str(Decimal(m.group(1))/1000)
     return ret_val
 
 def parse_sunos(monitor, output):
@@ -59,7 +61,7 @@ def parse_sunos(monitor, output):
     for row in output.split('\n'):
         m = re.search('^round-trip \(ms\)  min/avg/max/stddev = (?P<min>[0-9]+(?:\.[0-9]+)?).*', row)
         if (m):
-            ret_val = str(float(m.group(1))/1000)
+            ret_val = str(Decimal(m.group(1))/1000)
     return ret_val
 
 PING = {
@@ -110,7 +112,7 @@ class ICMPMonitor(AbstractMonitor):
                     )
                 if item is None:
                     item_type = self.dao.item_dao\
-                        .get_itemtype_by_name(name='float')
+                        .get_itemtype_by_name(name='decimal')
 
                     item = self.dao.item_dao\
                         .create_item(
